move to shared: configuration.ts types
