+++ client/cody/src/configuration.ts
@@ -1,16 +1,6 @@
 import * as vscode from 'vscode'
 
-export type ConfigurationUseContext = 'embeddings' | 'keyword' | 'none' | 'blended'
-
-export interface Configuration {
-    enabled: boolean
-    serverEndpoint: string
-    codebase?: string
-    debug: boolean
-    useContext: ConfigurationUseContext
-    experimentalSuggest: boolean
-    openaiKey: string | null
-}
+import type { ConfigurationUseContext, Configuration } from '@sourcegraph/cody-shared/src/configuration'
 
 export function getConfiguration(config: vscode.WorkspaceConfiguration): Configuration {
     return {
